refactor(state): extract file descriptor table into separate module
- Created src/state/ directory structure
- Moved FileDescriptor enum and FileDescriptorTable to src/state/fd_table.rs
- Added comprehensive module-level documentation for FD table
- Maintained backward compatibility through re-exports in src/state/mod.rs
- All 669 tests pass successfully
- Part of Phase 4 state.rs refactoring (step 4.1)
